Enable Dialog to close by pressing X button
Dialog button is now able to close when the X button is pressed. The X
button was not able to close the dialog before since the open prop of
the Dialog component was triggered only when the budget is created. To
create the dialog, user must press the "Create Budget" button